# week11-assignment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple example of using Gin with Swagger.",
        "title": "Simple API Example",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {"/books": {
            	"get": { "description": "Get details of book",
				"produces": ["application/json"],
                "tags": ["Books"],
                "summary": "Get all book",
                "responses": {"200": {"description": "OK",
				"schema": {"type": "array",
                    "items": {"$ref": "#/definitions/main.Book"}
                        }
                    },
                "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            	"post": {"description": "Create a new book",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Books"],
                "summary": "Create a book",
                "parameters": [
                    {
                        "description": "Book payload",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        		"/books/{id}": {"get": {"description": "Get details of a Book by ID",
                "produces": ["application/json"],
                "tags": ["Books"],
                "summary": "Get Book by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": { "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
           		"put": {"description": "Update an existing book by ID",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Books"],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Book payload",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    }
                ],
                "responses": {"200": {"description": "OK",
                "schema": {"$ref": "#/definitions/main.Book"}
             },
                "400": {"description": "Bad Request",
                    "schema": {"$ref": "#/definitions/main.ErrorResponse"}
            },
                "404": {"description": "Not Found",
                    "schema": {"$ref": "#/definitions/main.ErrorResponse"}
             },
                "500": {"description": "Internal Server Error",
                    "schema": {"$ref": "#/definitions/main.ErrorResponse"}
                    }
                }
            },
            "delete": {
                "description": "Delete a book by ID",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Books"],
                "summary": "Delete a book",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: book deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {"$ref": "#/definitions/main.ErrorResponse"}
                    }
                }
            }
        }
    },
    			"definitions": {"main.Book": {"type": "object",
            	"properties": { "author": { "type": "string"},
                "created_at": { "type": "string"},
                "id": {"type": "integer"},
                "isbn": {"type": "string"},
                "price": {"type": "number"},
                "title": {"type": "string"},
                "updated_at": {"type": "string"},
                "year": {"type": "integer"}
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}